requirements.txt
fastapi
uvicorn
openai
google-auth
google-auth-oauthlib
google-api-python-client
requests
python-dotenv
pydantic
python-multipart